+++ src/com/jordanbunke/stipple_effect/tools/BoxSelect.java
@@ -6,16 +6,17 @@ import com.jordanbunke.delta_time.utility.Coord2D;
 import com.jordanbunke.stipple_effect.project.SEContext;
 import com.jordanbunke.stipple_effect.selection.SelectionUtils;
 import com.jordanbunke.stipple_effect.utility.Constants;
+import com.jordanbunke.stipple_effect.utility.Settings;
 import com.jordanbunke.stipple_effect.visual.GraphicsUtils;
 import com.jordanbunke.stipple_effect.visual.SECursor;
 
 import java.util.HashSet;
 import java.util.Set;
 
-public final class BoxSelect extends ToolWithMode implements OverlayTool {
+public final class BoxSelect extends ToolWithMode implements OverlayTool, SnappableTool {
     private static final BoxSelect INSTANCE;
 
-    private boolean drawing;
+    private boolean drawing, snap;
     private Coord2D pivotTP, endTP, topLeft, bottomRight;
     private GameImage selectionOverlay;
 
@@ -25,6 +26,7 @@ public final class BoxSelect extends ToolWithMode implements OverlayTool {
 
     private BoxSelect() {
         drawing = false;
+        snap = false;
 
         pivotTP = Constants.NO_VALID_TARGET;
         endTP = Constants.NO_VALID_TARGET;
@@ -45,12 +47,14 @@ public final class BoxSelect extends ToolWithMode implements OverlayTool {
 
     @Override
     public String getCursorCode() {
-        return drawing ? SECursor.RETICLE : super.getCursorCode();
+        return drawing ? SECursor.RETICLE
+                : convertNameToFilename() + "_" + getMode().name().toLowerCase();
     }
 
     @Override
     public void onMouseDown(final SEContext context, final GameMouseEvent me) {
-        final Coord2D tp = context.getTargetPixel();
+        final Coord2D tp = snapToClosestGridPosition(
+                context, context.getTargetPixel());
 
         // this condition instead of constants.isTargetingPixelOnCanvas()
         // so that dragging can start off canvas
@@ -60,14 +64,15 @@ public final class BoxSelect extends ToolWithMode implements OverlayTool {
             if (ToolWithMode.getMode() == Mode.SINGLE)
                 context.deselect(false);
 
+            final Set<Coord2D> bounds = new HashSet<>(Set.of(tp));
+
             pivotTP = tp;
             endTP = tp;
             topLeft = tp;
-            bottomRight = tp;
+            bottomRight = SelectionUtils.bottomRight(bounds);
 
             selectionOverlay = GraphicsUtils.drawSelectionOverlay(
-                    context.renderInfo.getZoomFactor(),
-                    new HashSet<>(Set.of(tp)), false, false);
+                    context.renderInfo.getZoomFactor(), bounds, false, false);
         }
     }
 
@@ -83,9 +88,12 @@ public final class BoxSelect extends ToolWithMode implements OverlayTool {
 
             final Set<Coord2D> bounds = new HashSet<>(pivotTP.equals(endTP)
                     ? Set.of(pivotTP) : Set.of(pivotTP, endTP));
-            topLeft = SelectionUtils.topLeft(bounds);
-            bottomRight = SelectionUtils.bottomRight(bounds);
+            topLeft = snapToClosestGridPosition(context,
+                    SelectionUtils.topLeft(bounds));
+            bottomRight = snapToClosestGridPosition(context,
+                    SelectionUtils.bottomRight(bounds));
 
+            bounds.clear();
             for (int x = topLeft.x; x < bottomRight.x; x++)
                 for (int y = topLeft.y; y < bottomRight.y; y++)
                     bounds.add(new Coord2D(x, y));
@@ -101,9 +109,6 @@ public final class BoxSelect extends ToolWithMode implements OverlayTool {
             drawing = false;
             me.markAsProcessed();
 
-            if (pivotTP.equals(endTP))
-                return;
-
             final int w = context.getState().getImageWidth(),
                     h = context.getState().getImageHeight();
 
@@ -116,6 +121,25 @@ public final class BoxSelect extends ToolWithMode implements OverlayTool {
 
             context.editSelection(box, true);
         }
+    }
+
+    private Coord2D snapToClosestGridPosition(
+            final SEContext context, final Coord2D pos
+    ) {
+        if (!(isSnap() && canSnap(context)))
+            return pos;
+
+        final int pgX = Settings.getPixelGridXPixels(),
+                pgY = Settings.getPixelGridYPixels(),
+                x = (int) Math.round(pos.x / (double) pgX),
+                y = (int) Math.round(pos.y / (double) pgY);
+
+        return new Coord2D(x * pgX, y * pgY);
+    }
+
+    private boolean canSnap(final SEContext context) {
+        return context.renderInfo.isPixelGridOn() &&
+                context.couldRenderPixelGrid();
     }
 
     @Override
@@ -131,5 +155,15 @@ public final class BoxSelect extends ToolWithMode implements OverlayTool {
     @Override
     public boolean isDrawing() {
         return drawing;
+    }
+
+    @Override
+    public void setSnap(boolean snap) {
+        this.snap = snap;
+    }
+
+    @Override
+    public boolean isSnap() {
+        return snap;
     }
 }